* tp/Texinfo/XS/convert/converter.c (retrieve_converter) (unregister_converter_descriptor), tp/Texinfo/XS/convert/get_converter_perl_info.c (get_sv_converter), tp/Texinfo/XS/main/build_perl_info.c: use size_t for converter descriptor consistently.
* tp/Texinfo/XS/convert/ConvertXS.xs (html_get_target),
tp/Texinfo/XS/convert/convert_html.c (prepare_special_units)
(html_prepare_conversion_units)
(html_prepare_output_units_global_targets)
(html_prepare_units_directions_files),
tp/Texinfo/XS/convert/get_html_perl_info.c (html_find_element_from_sv)
(get_output_units_descriptor_converter_sv)
(element_converter_from_sv), tp/Texinfo/XS/main/build_perl_info.c,
tp/Texinfo/XS/main/converter_types.h (CONVERTER),
tp/Texinfo/XS/main/get_perl_info.c (get_sv_output_units_descriptor)
(get_sv_output_units, find_root_command, find_element_from_sv),
tp/Texinfo/XS/main/output_unit.c (retrieve_output_units)
(split_by_node, split_by_section),
tp/Texinfo/XS/structuring_transfo/StructuringTransfoXS.xs
(split_by_node, split_by_section, rebuild_output_units): use size_t
for output units descriptor consistently.

* tp/Texinfo/XS/convert/converter.c (new_converter, id_to_filename)
(free_comma_index_subentries_tree, find_output_unit_file)
(free_output_unit_files_file), tp/Texinfo/XS/main/output_unit.c
(unsplit, split_pages, units_directions, units_file_directions): use
size_t instead of int where the variable could reach a value outside
of the int range and also when the index is better represented by a
size_t.  Cast to/from int and size_t when needed.

+++ tp/Texinfo/XS/convert/convert_html.c
@@ -1564,15 +1564,16 @@ compare_special_units (const void *a, const void *b)
 }
 
 void
-prepare_special_units (CONVERTER *self, int output_units_descriptor)
+prepare_special_units (CONVERTER *self, size_t output_units_descriptor)
 {
   size_t i;
   STRING_LIST *special_unit_varieties = &self->special_unit_varieties;
   SPECIAL_UNIT_ORDER *special_units_order;
   OUTPUT_UNIT *previous_output_unit = 0;
 
-  int special_units_descriptor = new_output_units_descriptor (self->document);
-  int associated_special_units_descriptor
+  size_t special_units_descriptor
+         = new_output_units_descriptor (self->document);
+  size_t associated_special_units_descriptor
          = new_output_units_descriptor (self->document);
 
   /* retrieve after reallocating */
@@ -1832,7 +1833,7 @@ static const enum command_id conf_for_documentlanguage[]
 void
 html_prepare_conversion_units (CONVERTER *self)
 {
-  int output_units_descriptor;
+  size_t output_units_descriptor;
 
   if (self->conf->USE_NODES.o.integer > 0)
     output_units_descriptor = split_by_node (self->document);
@@ -5442,7 +5443,7 @@ html_prepare_output_units_global_targets (CONVERTER *self)
   const OUTPUT_UNIT *top_output_unit = get_top_unit (self->document,
                                                      output_units);
 
-  int special_output_units_lists[2] = {
+  size_t special_output_units_lists[2] = {
     self->output_units_descriptors[OUDT_special_units],
     self->output_units_descriptors[OUDT_associated_special_units]};
 
@@ -5529,7 +5530,7 @@ html_prepare_output_units_global_targets (CONVERTER *self)
      directions */
   for (i = 0; i < 2; i++)
     {
-      int special_units_descriptor = special_output_units_lists[i];
+      size_t special_units_descriptor = special_output_units_lists[i];
       const OUTPUT_UNIT_LIST *units_list
        = retrieve_output_units (self->document, special_units_descriptor);
       if (units_list && units_list->number)
@@ -5544,7 +5545,7 @@ html_prepare_output_units_global_targets (CONVERTER *self)
   s = 0;
   for (i = 0; i < 2; i++)
     {
-      int special_units_descriptor = special_output_units_lists[i];
+      size_t special_units_descriptor = special_output_units_lists[i];
       const OUTPUT_UNIT_LIST *units_list
         = retrieve_output_units (self->document, special_units_descriptor);
       if (units_list && units_list->number)
@@ -6291,7 +6292,7 @@ html_prepare_units_directions_files (CONVERTER *self,
 {
   size_t i;
   FILE_SOURCE_INFO_LIST *files_source_info = 0;
-  int external_nodes_units_descriptor
+  size_t external_nodes_units_descriptor
      = new_output_units_descriptor (self->document);
   OUTPUT_UNIT_LIST *external_node_target_units
     = retrieve_output_units (self->document, external_nodes_units_descriptor);
